Refactor: put HRRR functions in designated module
